Update patch set 12
Patch Set 12:

(1 comment)

> Ping. Happy to take this over and get it merged if you want?

Hi,

Thanks for the ping.

I did plan to continue with this during the next two weeks, but if you can continue with it right now, you're welcome.

The current patch is very long, but almost all of it is quite trivial—mostly moving the same code around to different files. The biggest part that is left is the ChemPage, which will require a bit more thought.

Patch-set: 12
Reviewer: Amire80 <54@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0