Purge formulae on action=purge
Previously, the 'action=purge' parameter could not be
accessed by the math extension, since 'action=purge' caused a reload
of the page, even before the math extension was called. However,
before MediaWiki 1.29.+ additional GET request parameter were
forwarded to the next page and could be evaluated by the math
extension. Thus, an additional parameter 'mathpurge=true' was added
to re-render formulae. However, it seems that this additional GET
parameter is no longer being forwarded.

This change reads the refer header to check if 'action=purge' was
passed. This makes the mathpurge parameter obsolete.

Bug: T74537
Change-Id: Ia09c039de3a7232a28bfdd5a97daf1029e5d7c40