Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

I like the update to the new shell API. However, please note that the MathTexvc Class is deprecated and will be removed very soon. With regard to the other changes I think the Math extension is not the correct place to change the behavior.
- Configuring the path environment is something that should not be done in the math extension.
- As stated in the commit message stdout should always be tracked in debug mode. So this is something that should be changed in mediawiki core.

Patch-set: 1
Label: Code-Review=-1