Update patch set 1
Patch Set 1:

(1 comment)

Agreeing that the stderr output change should be part of the Mediawiki shell api, gonna create a new patch for that when I find the time. Was aware this class is deprecated but as this issue was quite painful to resolve I wanted to spare other users the same until it is actually removed. Feel free to close.

Patch-set: 1
Label: Verified=0