Update patch set 3
Patch Set 3:

I'd argue that using `$wgTexvc` only could also be worse in terms of intuitive use. For example we currently don't have `$wgTexvc` set in our config as we use the expected default location.

Passing the `PATH` via `$wgTexvc` would certainly be possible with a little further refactoring, exploding `$wgTexvc` before adding it to `$cmdArgs` should do the trick basically. Though as soon as spaces in paths need to be considered this becomes very brittle or needlessly complex. It would also make meaningful error reporting harder as we need to either parse `$wgTexvc` or stderr output at all times. The user does get a good error code now, adding that to the troubleshooting section of the docs, along with some examples, should be fine IMO. The current error log message when no `PATH` is set could also be expanded upon for variable usage.

Fixing this in the `texvc` binary is not an option, it works as it should in that regard, anything else would be bad practice.

Patch-set: 3
Reviewer: Dannenberge <5825@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0