Update patch set 3
Patch Set 3:

I really do not know what "info in main package" refers to. If you want to do it differently than this patch proposed then please feel free to do so. But based on what I see in the history of this codebase I can not support such a change. I'm also not a lawyer.

Patch-set: 3