Update patch set 3
Patch Set 3:

If we use the case examples from WordPress, then by extension it is a license violation for any MediaWiki extension to use a license not compatible with core given the standalone creation of extensions has been interpreted by some as a derivative work. (I'd disagree, but I'm not a lawyer). Anyway, it's imho fine to correct Math  to GPLv2+ as GPLv2-only was *never* specified here. It also didn't specify "+" or "later" until recently, but then again, many repos didn't, it was just an oversight. 

Anyway, if we want to go for relicense and/or confirmation, we'll have to contact those who's name is listed in various copyright clauses. Currently: Tomasz Wegrzanowski, Brion Vibber, Moritz Schubotz and Derk-Jan Hartman.

Patch-set: 3
Label: Code-Review=0