Add fields to composer.json required for install
Certain basic fields (primarily "name") are required when installing via
composer (through packagist or as a composer vcs repository). Although
composer is not the official means of extension installation for
MediaWiki, it is being experimented with in the context of the WMF
Continuous Delivery pipeline. Adding such fields is harmless and allows
such experimentation to move forward.

Change-Id: I7b8583aa15131417e2701ec2d82f6cffe028332d