Update patch set 2
Patch Set 2:

> It's an experiment and we just need it for Math at the moment

We have been down this road before, e.g. in WikiEditor :) The problem is that if this gets merged, it'll end up on packagist, people will start depending upon it, and then whenever we try undoing the experiment, people will inevitably complain that we broke their thing (and we should avoid that unless necessary). I'll suggest alternatives on the task.

Patch-set: 2