Update patch set 3
Patch Set 3:

I see the code removed here is not used anywhere any more. So literally dead code.

From what I understood the same is true for the "math" database table. Isn't it? It was already unused before this patch. So any third-party who is running the Math extension in this very version already has a dead table in the database. Why not remove it then?

TL;DR: I think this patch needs to contain the update script to remove the unused table from old instances. This update script can easily be added in a later patch (I would even suggest to do this in a later, separate patch). But I would like to see at least a draft for this patch first before merging this, so it will not be forgotten.

Apart from this I don't see an issue here and am willing to merge this.

Patch-set: 3
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0