formatter-factory-callback should not load math styles
Currently making use of Wikidata functions has the side effect
of pulling in styles and scripts that are not needed by the currently
rendered page.

It's not really clear why this needs to do this. My albeit weak
understanding of this code is that onWikibaseRepoDataTypes
should take care of loading math styles inside instances of Wikibase
but the use of this hook inside a wiki relying on Wikibase does not
need the styles - this should be added elsewhere.

I suggest removing this code unless someone can explain to me why
this is necessary (in which case I'll investigate this bug further)
but right now this looks redundant to me.

Bug: T173949
Change-Id: I6c17c687c7c16504e4929bb6630740a754a401fb