Add Math styles and scripts in BeforePageDisplay hook
formatter-factory-callback should not load math styles

Currently making use of Wikidata functions has the side effect
of pulling in styles and scripts that are not needed by the currently
rendered page.

Modules that relate to stuff on the page inserted by the power of wikitext,
should be associated with the ParserOutput object. Using $wgOut can cause
unexpected side effects such as parser cache pollution (which is what
we are seeing in mobile)

The "elsewhere" I propose is inside BeforePageDisplay - as this mirrors
how we add scripts and styles for other optional components inside
Skin::getDefaultModules and provides direct access to OutputPage (no
need to use a global). Given this also runs at the end of rendering
we can be safely assured that the HTML markup will not change any
further.

Bug: T173949
Change-Id: I6c17c687c7c16504e4929bb6630740a754a401fb