Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

I'm not so sure this is an issue that needs fixing. What exactly are you trying to solve here? Did you run into an actual issue with this?

Note that just removing this will cause issues in situations that rely on these styles and scripts. Therefor the -1. If this is the case depends on the Math renderer you picked in your personal settings: https://en.wikipedia.org/wiki/Special:Preferences#mw-prefsection-rendering

Don't get me wrong, I pretty much agree with the analysis outlined in the commit message. The code appears misplaced. But this is mostly because the signature of the hook does not contain an OutputPage, Request, or Parser object. This is why the hook handler currently accesses some global that might or might not be related to the OutputPage the Wikibase entity containing a Math value is going to show up.

An easy improvement is to move the code up a level, and run it directly in the onWikibaseRepoDataTypes and onWikibaseClientDataTypes hook handler functions.

A better place for this code is in a hook that has access to:
* the OutputPage
* the namespace
* the users settings

This code should then check NamespaceChecker::isWikibaseEnabled() and only load the .css and .js in Wikibase namespaces. A typical place for this is in a BeforePageDisplay hook handler.

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1