Update patch set 1
Patch Set 1:

> I'm not so sure this is an issue that needs fixing. What exactly are you trying to solve here? Did you run into an actual issue with this
All the details are in the bug report? Ive spent several days investigating a parser cache bug that this has been causing for over a year now. Open to other solutions but i sadly dont understand enough of the wikibase and math code to know why styles and scripts need to be loaded at this point in the code... i agree with you they can probably be removed in both places but this particular hook is the one im seeing issues with. They are already added in hooks elsewhere in the code. I'm  not sure why they would be added here too...

Patch-set: 1
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>