Update patch set 3
Patch Set 3:

Just had a chat with Physikerwelt. We realized the appropriate replacement for the code in question is something like it is done in I3da2920. Essentially a dumbed-down version of https://gerrit.wikimedia.org/r/#/c/mediawiki/extensions/Wikibase/+/434177/2/repo/includes/ParserOutput/GlobeCoordinateKartographerDataUpdater.php

The idea is to have a MathDataUpdater registered in Wikibase. This DataUpdater gets a chance to see every statement on an entity page. This way it can check if a MathValue appears somewhere on the page, and memorize this in a boolean flag. At the end, updateParserOutput() is called one time, and the DataUpdater can load the two .css and .js modules, if the flag is true.

The only issue I see is that there seems to be no entry point for extensions to register their own DataUpdaters. But this can be added.

Patch-set: 3
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>