Update patch set 4
Patch Set 4:

> Is that acceptable?

From my perspective, yes.

Even if this patch here introduces a hack that is quite problematic for a series of reasons (a possibly expensive substring scan on *every* page for a substring that might or might not suficently detect the use of MathValues on entity pages), I believe it is better than the previous hack.

But I can not make the final call on this, as I don't have the capacity to test this with all of the Math renderers.

One improvement I would love to see is to limited this code to namespaces where Wikibase is enabled.

Patch-set: 4
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>