Update patch set 4
Patch Set 4: Code-Review-1

I disagree on it not being a problem. The Skin "solution" is also a problem. It has been tolerated there because those hacks in question relate to loading of styles for components not natively recognised by the wiki, in other words, there is no callback for adding them, because they are part of wikitext (they = sortable tables, and mw-ui-botton).

The Math content, however, seems very specifically as something that doesn't just appear on the page through basic wikitext. They are styles and scripts for elements produced by the Math extension itself.

How are those added to the output? It seems unlikely that there is no Parser object available where that happens, given they are inserted by the parser, right?

Also, aside from it seeming like tech debt (per the above, this should have a callback in the parser), and it being inefficient (per Thiemo), it would also regress us with regards to T140664 given that this BeforePageDisplay hook only applies to page views, not to other ways of obtaining page content. E.g. ApiParse, previews, Parsoid, VE, would not have the styles.

So if we take the debt and inefficiency (and plan to repay it next quarter?), we'd at least have to find a way to do from a hook that applies to the parser or skin as well.

Patch-set: 4
Reviewer: Krinkle <34@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1