Remove unnecessary addModule and addModuleStyle calls
From what I can see the mathTagHook is the only place where
Math elements can be rendered and thus the only place we need
to add modules and styles. These two can be safely removed.

The Math styles and scripts modules are are only needed if Math
is rendered and from looking at the code the mathTagHook is the only
way to do that.

Note: also adding them here has unwanted side effects of parser cache
pollution

Bug: T173949
Change-Id: I0abd6acbfa12fd503d11476f0c8c8c8a8a851cdc