Update patch set 4
Patch Set 4:

> The Math content, however, seems very specifically as something that doesn't just appear on the page through basic wikitext. > They are styles and scripts for elements produced by the Math extension itself.

> How are those added to the output? It seems unlikely that there is no Parser object available where that happens, given they > are inserted by the parser, right?

That's what I originally thought. An alternative solution would be:
https://gerrit.wikimedia.org/r/#/c/mediawiki/extensions/Math/+/442725/
This however was rejected and I'm still unclear why. I'm hoping it was just a misunderstanding as it doesn't prevent the loading of math styles and scripts from what I can see given they are already being added elsewhere.

It sounds like this problematic code loads Math styles on Wikidata entities when certain statements are loaded (but I'm not clear which). If so as a short term measure can we limit it to Wikidata or to a JS load? (see Thiemo Kreuz (WMDE) comment there)

If neither of these solutions works, there is a third option - to disable $wgMathEnableWikibaseDataType on mobile views to short circuit this code.

Looking for **some** short term solution here. I don't mind what, but it's of high importance to my team to get this fixed, given some changes we are planning, but it seems like we don't have the power/enough knowledge to fix it.

Patch-set: 4
Reviewer: Jdlrobson <94@e9e9afe9-4712-486d-8885-f54b72dd1951>