Update patch set 1
Patch Set 1:

> I am getting the impression that this is "state-of-the-art" [1]. However, given all the discussions we had in the past on sustainable coding I get the feeling that this exactly what we wanted to avoid:-( 

In my mind, it's both: a) the best approach for now and b) technical debt that should be addressed on the Wikibase side by introducing a proper extension point. As the maintainer of the Math extension, it's ok for you to go with (a) and not worry about (b). But if you want to take on (b), great! Coordinate with the Wikidata team and go for it!

Patch-set: 1
Reviewer: Daniel Kinzler <128@e9e9afe9-4712-486d-8885-f54b72dd1951>