Update patch set 2
Patch Set 2:

The error is unrelated to this change. The +2 is set, and will not go away. But I will not manually merge this without the error being fixed first, as there is no need to do so. This change can easily wait till that happened. It won't get lost.

The Math or Wikibase team (which is where the error probably comes from) will look into the error.

Patch-set: 2
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>