Update patch set 2
Patch Set 2:

> It looks like state of the art is to *not* do this. Instead,
 > \DatabaseUpdater::addExtensionTable takes care if the table exists
 > or not.
 > 
 > What is the issue you want to fix here? Is there a Phabricator
 > ticket describing the problem?

Currently there’s no bug filled for this because it’s not a bug. Rather it’s how the extension author does it (checks the table). There’s no problem with this extension but with other extensions you can find multiple create table statements and then only checking one of the tables. So if the table that is check does not exists then the sql is executed thus you have a problem where it try’s to create a already existent table thus fails.

Patch-set: 2
Reviewer: Paladox <1665@e9e9afe9-4712-486d-8885-f54b72dd1951>