Update patch set 9
Patch Set 9: Code-Review+1

(2 comments)

My previous +1 holds. I think the code is fine, and I'm willing to merge it, if nobody else does it. The only thing I find "not good enough" (TM) yet is the name of the category. Changing it later might be more of a hassle than changing it right now. But this is close to bike-shedding ;-) and only meant as a suggestion. I can merge it as it is if you prefer.

Patch-set: 9
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1