Also logging the STDERR message
Anytime, when execute mathoid/cli.js failed, only logging Mathoid cli mathoid/cli.js is not executable message. This will make operator messy, because the real reason is often not that the file is not executable, such as Nodejs throw an error or failed to execute shell command.

Logging error messages from STDERR at the same time would be better.

Bug: T198564
Change-Id: I7d0ae0f3ab4789124f9ff1533e7712ca31233f9e