Update patch set 1
Patch Set 1: Code-Review-2

If the script is incompatible to the new API this can should be fixable. The purpose of this script is to generate test cases. One might discuss to move the script to another repo, but I think here it's easy to discover.

Patch-set: 1
Reviewer: Physikerwelt <183@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-2