Update patch set 1
Patch Set 1:

Alright, so this is not a false positive. printSvgError calls htmlspecialchars on its argument. The argument here is MathRenderer->getLastError(), which returns $this->lastError, which is MathInputCheckRestbase::getError(), which takes restbaseInterface->getError() (that seems to be a message key), and calls errorObjectToHtml on it, which in turn calls escaped().

As you can see, this is a pretty long chain. Now, I don't know if it would make sense to change errorObjectToHtml to call text(), but given that the method is called "...ToHtml", I'd leave it return something HTML-safe. Plus we'd risk breaking something else relying on the HTML safety of the return value. So the alternative is to remove the escaping from printSvgError. Given that there are only 3 uses of that method, and two of them with a literal, this is what I'm going to do.

Patch-set: 1