Add popup support for annotated math
Add a special page and an API endpoint to fetch data from Wikidata
items with a given qID. The special page summarizes information from
Wikidata. The API endpoint allows to request the information
directly. Both, the API endpoint and the special page, fetch the data
from a new helper class for consistency.

Bug: T208758
Bug: T229939
Change-Id: Idd22057a88312bf1a1cb5546d0a6edca5678d80d