Added support for popups for annotated math
Added a special page and a API endpoint to fetch data from wikidata
items with a given qID. The special page summarizes information from
wikidata. The API endpoint allows to request the information
directly. Both, the API endpoint and the SpecialPage, fetch the data
from a new helper class for consistency. The Popups extension was
extended to fetch data from wikidata with this new API endpoint.

Patch 4:
-> Updated phan config
-> Changed to a better OOUI layout for the special page
-> Added some error handling

Bug: T208758
Change-Id: Idd22057a88312bf1a1cb5546d0a6edca5678d80d