Update patch set 10
Patch Set 10: Code-Review-1

(1 comment)

The concepts "Wikidata" and "Wikibase" seem to be used interchangeably in this PS, but they are not the same thing and a clear distinction should be made. Also, this PS assumes that it will be deployed only in WMF production (where, effectively Wikidata === Wikibase), but that's not the case. There are other WB installations out there that are not called Wikidata.

This patch ads a global query module named wikidataextracts. Since it resides in the Math extension, this is not acceptable. Furthermore, the module seems to be doing generic WB item lookups, something that is already achievable via the standard WB API modules. Therefore, this PS should be reworked so that it makes use of existing APIs rather than adding redundant modules for the convenience of the Math extension.

Patch-set: 10
Reviewer: Mobrovac <2177@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1