Update patch set 10
Patch Set 10:

Correct me if I'm wrong, but here you want to get info about WD items, right? What I specifically mean is that you are really interested in concrete Wikidata items and terms (as opposed to Wikibase's general concept of items and terms). In other words, you are only interested in what is available on wikidata.org. If that is the case, then the best way to go about this is to directly place web requests to wikidata's api available online. Mediawiki provides the HTTP class which can be used to that end.

As for which Wikidata API module to call in order to get the info you need, please take a look at their API documentation~[1,2] which lists all the available modules.

[1] https://www.mediawiki.org/wiki/Wikibase/API
[2] https://www.wikidata.org/w/api.php

Patch-set: 10
Reviewer: Mobrovac <2177@e9e9afe9-4712-486d-8885-f54b72dd1951>