Update patch set 16
Patch Set 16: Verified-1

I assume the current code implements the screenshot in T208758. I have a product concern about this: We intentionally do not expose any Wikidata IDs to Wikipedia readers (and try very hard to avoid it for editors). We can not assume that the reader knows anything about Wikidata much less what to do if they're given a Wikidata ID - or even that it is a Wikidata ID. We also can't rely on Qsomenumber always indicating Wikidata as other Wikibase installations will also have the same IDs.
So my recommendation from a product-side is to remove any mention of Wikidata and Wikidata IDs from the popup. That way it will be just as informative and less confusing for readers.

Patch-set: 16
Reviewer: Lydia Pintscher <176@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=-1