Update patch set 11
Patch Set 11: Code-Review-1

(3 comments)

I'm afraid this doesn't look like proper error handling to me. If the input is that much broken, it should not be used, but report an error instead. It might be that a few of these properties are optional, and the rest of the code would continue to work just fine without them. But I'm not an active contributor to this codebase and can't tell. With the limited knowledge I have I must assume all properties are critical.

Setting a -1 to make my concern more visible. Please feel free to ignore my vote in case I'm wrong.

Patch-set: 11
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1