Update patch set 2
Patch Set 2:

The basic idea goes like this: What is the main purpose of an abstract class? It's when multiple classes need to share some pieces of code that are the same in all classes. In other words: the only reason to make something abstract instead of an interface is code-sharing.

This type of code-sharing is never a good thing. For example, subclasses can still override methods already being implemented in the abstract base class, which makes the thing rather pointless. Subclasses are also not forced to call the code from the base class. And what happens if you actually want to change the implementation in the base class? How do you make sure this only affects the one subclass you care about, but not the others? Testing this is hard.

In other words: abstract classes don't give any additional guarantee. All they do is providing a way to avoid copy-pasting some code.

A much better technique is to have the code in a small, isolated service class that can be tested separately, and injected into classes that actually need to call the service.

Providing common guarantees should be done via interfaces, not via abstract classes.

As always, there are exceptions.

Patch-set: 2
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>