Update patch set 2
Patch Set 2:

(1 comment)

I disagree with this being a reason for a -1. There is nothing wrong with this patch. The fact we can continue to improve the code can't be a reason to block a patch. Especially because *removing* a method would be out of the scope of this patch here, which is about updating the CodeSniffer.

Please upload a follow-up patch, or update this one here if you prefer. Please update the commit message as well then. Thank you.

Patch-set: 2
CC: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>