Update patch set 3
Patch Set 3: -Code-Review

The problem with the error message is that it will be logged for each formula. If people did not install logrotate it will fill all the disk space.
It would be better if mathoid would expose its capabilities upfront and this information would be displayed on Special:MathStatus.

Patch-set: 3
Label: Code-Review=0