Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

This is great. One comment inlined. 

You can also tag T187154 as it's a step towards it. If you have time and energy, it would be amazing to namespace all the classes in the extension - the would need to be done eventually. If you are only interested in this as preparation for the REST API stuff - fine with me as well. Up to you.

- Only -1 for the inlined comment, not requiring to namespace everything.
- Adding Reedy for visibility

Patch-set: 1
Reviewer: Reedy <3@e9e9afe9-4712-486d-8885-f54b72dd1951>
Reviewer: Ppchelko <2681@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-1