Update patch set 8
Patch Set 8:

Ok. So when depending on wikibase in CI, we actually also depend on all the transitive dependencies. Thus, Wikibase depends on CirrusSearch, and we end up installing it. I suspect that to be a reason search tests fail.

@Physikerwelt - why is Math dependent on Wikibase in CI?

Patch-set: 8
Reviewer: Ppchelko <2681@e9e9afe9-4712-486d-8885-f54b72dd1951>