Add mathoid formula and render REST endpoint
Expose the math formula and render service via the REST
API.

Port the behaviour of the RESTbase endpoints to a new
endpoint that uses the MediaWiki API.

Since the new endpoints depend on WAN cache the cache
must not be set to NO_CACHE.
Otherwise, the endpoints will certainly not function.
This is worth noting since WAN cache defaults to NO_CACHE.
If the cache is ever cleared the x-resource-location
is useless, as the original input can not be generated from
the hash.

For the PNG images, we use a StringStream since the image
data is stored in an array and does not originate from a
file-system stream.

Tests are provided as API tests, but are not yet enabled
in the CI pipeline.

Establishing the mathoid backend as new service does not
very well align with the functional grouping of the
checking service. However, the differences between
mathoid and other rendering backends, such as LaTeXML
are quite huge. Thus a functional abstraction for
redering seems infeasible at this point.

Bug: T255807, T255808
Change-Id: I47d201d65e2064d21cc5d311ffe3018c5fc77ce8