MathStatus: Restore logged-in restriction, and make unlisted
== Background ==

This follows-up I98eabfbeda2fbef3c (T252362) which replaced an
inline "purge" permission check, with a declarative "purge"
restriction in 2020. This way MediaWiki knows about it, and thus it becomes
effectivey unlisted from Special:SpecialPages, when viewing it
as logged-out user.

In 2023, with MediaWiki 1.41, the "purge" right was replaced with
an implicit grant that both logged-out and logged-in users have, thus
making Special:MathStatus listed Special:SpecialPages, and this time
world-accessible.

== This change ==

* Make the page explicitly unlisted, instead of using a user right
  as proxy for this.

* Make the page explicitly for logged-in users only,
  like it was prior to I98eabfbeda2fbef3c, but based on isNamed
  instead of by proxy of a user right.

Bug: T344206
Change-Id: I408d1d71dc8cb3c32ad6bd124726d96b26209a1b